Extracted stars to it's own component
