+++ src/components/products/ProductCard/ProductCard.css
@@ -19,7 +19,11 @@
     display: inline-block;
 }
 
-.card a {
+.card-body a {
     text-decoration: none;
     color: black;
+}
+
+.card-body a:hover {
+    color: rgb(0, 0, 255);
 }
